refactor: consolidate location APIs into `LocationManagerCompat`
