+++ app/build.gradle
@@ -223,9 +223,6 @@ dependencies {
     // For UART access
     implementation 'com.github.mik3y:usb-serial-for-android:3.7.0'
 
-    // location services
-    googleImplementation 'com.google.android.gms:play-services-location:21.0.1'
-
     // For Firebase Crashlytics & Analytics
     googleImplementation platform('com.google.firebase:firebase-bom:32.7.0')
     googleImplementation 'com.google.firebase:firebase-crashlytics-ktx'
