+++ app/src/main/java/com/geeksville/mesh/repository/location/LocationRepositoryModule.kt
@@ -1,23 +1,20 @@
 package com.geeksville.mesh.repository.location
 
 import android.content.Context
+import android.location.LocationManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.GlobalScope
 import javax.inject.Singleton
 
-@OptIn(DelicateCoroutinesApi::class)
 @Module
 @InstallIn(SingletonComponent::class)
 object LocationRepositoryModule {
 
     @Provides
     @Singleton
-    fun provideSharedLocationManager(
-        @ApplicationContext context: Context
-    ): SharedLocationManager = SharedLocationManager(context, GlobalScope)
+    fun provideLocationManager(@ApplicationContext context: Context): LocationManager =
+        context.applicationContext.getSystemService(Context.LOCATION_SERVICE) as LocationManager
 }
